+++ src/main/java/techcourse/myblog/web/ArticleController.java
@@ -32,19 +32,19 @@ public class ArticleController {
 	}
 
 	@GetMapping("articles/{articleId}")
-	public String getArticle(@PathVariable int articleId, Model model) {
+	public String getArticle(@PathVariable Long articleId, Model model) {
 		model.addAttribute("article", articleRepository.findById(articleId));
 		return "article";
 	}
 
 	@GetMapping("articles/{articleId}/edit")
-	public String editArticle(@PathVariable int articleId, Model model) {
+	public String editArticle(@PathVariable Long articleId, Model model) {
 		model.addAttribute("article", articleRepository.findById(articleId));
 		return "article-edit";
 	}
 
 	@PutMapping("articles/{articleId}")
-	public String modifyArticle(@PathVariable int articleId, Article article, Model model) {
+	public String modifyArticle(@PathVariable Long articleId, Article article, Model model) {
 		article.setId(articleId);
 		articleRepository.update(article);
 		model.addAttribute(articleRepository.findById(articleId));
@@ -52,7 +52,7 @@ public class ArticleController {
 	}
 
 	@DeleteMapping("articles/{articleId}")
-	public String deleteArticle(@PathVariable int articleId) {
+	public String deleteArticle(@PathVariable Long articleId) {
 		articleRepository.deleteById(articleId);
 		return "redirect:/";
 	}
